Got logging to work.
Decided that it is likely best for every battery to have their own logger rather than try to send data to some all-seeing 'log manager'. The data is right there, I just needed to send a message back through rabbitMQ on an exchange called the "log-flags" exchange, and now the data gets logged at the same time that it posts the data to the queue. Every battery will probably subscribe to that, and check the message for its own identifying info, and if it's there, it will toggle the loggle. I still need to implement:

--Data producers (aka batteries) comparing the "battery-id" string to their own identifying information, and if the info doesn't match up then just don't take action
--Need to format the data to .csv. This shouldn't be a lot of work, but it will be a little tedious and will involve f strings.
--The header needs to be printed to the .csv file every single time we enable logging.
--The frontend isn't pulling the data path from my selection yet, that needs to be fixed.